+++ TCS3472_ex.h
@@ -97,7 +97,7 @@ __INLINE FctERR INLINE__ TCS3472_Write_Ctl(const uint8_t ctl) {
 	return TCS3472_Write((uint8_t *) &ctl, TCS3472__CONTROL, 1); }
 
 
-/*!\brief oscillator Enable / Disable
+/*!\brief Oscillator Enable / Disable
 ** \param[in] en - 0 Disable, 1 Enable
 ** \return FctERR - error code
 **/